+++ src/lib.rs
@@ -1,29 +1,95 @@
-use pest::Parser;
+use pest::{iterators::Pairs, Parser};
 
 mod error;
 mod parser;
 mod rollresult;
 pub use error::*;
 
-use parser::{compute, RollParser, Rule};
+use parser::{RollParser, Rule};
 pub use rollresult::RollResult;
 
-/// Compute a roll expression
-pub fn roll(input: &str) -> Result<RollResult> {
-    let mut pairs = RollParser::parse(Rule::command, input)?;
-    let mut roll_res = compute(pairs.next().unwrap().into_inner())?;
-    if let Some(reason) = pairs.next() {
-        if reason.as_rule() == Rule::reason {
-            roll_res.add_reason(reason.as_str()[1..].trim().to_owned());
+const REASON_CHAR: char = ':';
+
+/// An object holding the query
+///
+/// It has no advantage compare to free function that would take `&str` as parameter (like previous
+/// version) but it provides a stable API for the day `pest` can have a `Send` type as the parse
+/// result.
+///
+/// see https://github.com/pest-parser/pest/issues/472
+/// and https://users.rust-lang.org/t/how-to-deal-with-external-type-which-is-send-and-sync/47530
+///
+#[derive(Clone)]
+pub struct Roller(String);
+
+impl Roller {
+    /// Store the input
+    ///
+    /// At version 1.0.0, it always returns Ok(Self).
+    ///
+    /// This is to have a stable API for further optimization where the parsing is done here (so it
+    /// can fail) and saved, see `Roller` documentation above.
+    ///
+    pub fn new(input: &str) -> Result<Self> {
+        Ok(Roller(input.to_owned()))
+    }
+
+    /// Evaluate and roll the dices
+    pub fn roll(&self) -> Result<RollResult> {
+        let mut pairs = RollParser::parse(Rule::command, &self.0)?;
+        let mut roll_res = parser::compute(pairs.next().unwrap().into_inner())?;
+        if let Some(reason) = pairs.next() {
+            if reason.as_rule() == Rule::reason {
+                roll_res.add_reason(reason.as_str()[1..].trim().to_owned());
+            }
+        }
+        Ok(roll_res)
+    }
+
+    /// Get an iterator on the dices in the expression
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use caith::Roller;
+    ///
+    /// let r = Roller::new("1d6 + 1d4 + 1d10 + 1d20").unwrap();
+    /// assert_eq!(vec!["1d6", "1d4", "1d10", "1d20"], r.dices().expect("Error on parse").collect::<Vec<_>>());
+    /// ```
+    pub fn dices(&self) -> Result<Dices> {
+        let pairs = RollParser::parse(Rule::command, &self.0)?
+            .next()
+            .unwrap()
+            .into_inner();
+        Ok(Dices { pairs })
+    }
+
+    /// Give back the query string
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+
+    pub fn trim_reason(&mut self) {
+        if let Some(idx) = self.0.find(REASON_CHAR) {
+            self.0 = self.0[..idx].to_owned()
         }
     }
-    Ok(roll_res)
 }
 
-/// Look for the first dice in expression and return it
-pub fn find_first_dice(input: &str) -> Result<String> {
-    let mut pairs = RollParser::parse(Rule::command, input)?;
-    parser::find_first_dice(pairs.next().unwrap().into_inner())
+/// Iterator that lazily returns each dice of the expression
+///
+/// See `Roller::dices()` for example
+///
+pub struct Dices<'a> {
+    pairs: Pairs<'a, Rule>,
+}
+
+impl<'a> Iterator for Dices<'a> {
+    type Item = String;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        parser::find_first_dice(&mut self.pairs)
+    }
 }
 
 #[cfg(test)]
@@ -32,6 +98,11 @@ mod tests {
 
     #[test]
     fn sandbox_test() {
-        eprintln!("{}", roll("4dF ! pouet").unwrap());
+        let r = Roller::new("1d6 + 1d4 + 1d8 + 1d10 * 1d20").unwrap();
+        r.dices()
+            .expect("Error while parsing")
+            .for_each(|d| eprintln!("{}", d));
+
+        eprintln!("{}, {}", r.as_str(), r.roll().unwrap());
     }
 }