+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+1.0.0
+- Add `Roller` to expose stable API for future optimization
+- BREAKING: Removed free function `roll` and `find_first_dice`, use `Roller` to get all the dices
+- BREAKING: reason starts with `:` instead of `!` 
+- `!` is now an alias for `ie`
+- on both exploding dice (`!` and `e`), if number is omitted, default to dice max
+- remove dep to `once_cell`
+
 0.5.0
 - Add `find_first_dice`: look for the first dice of the expression and return it